+++ src/app/features/events/events-routing.module.ts
@@ -6,7 +6,7 @@ import { AuthGuard } from '@pages/login/auth.guard';
 import { EventsComponent } from '@features/events/events.component';
 import { EventsFavComponent } from '@features/events/events-fav/events-fav.component';
 import { EventViewComponent } from '@features/events/event-view/event-view.component';
-import { EventAdminComponent } from '@features/events/event-admin/event-admin.component';
+import { EventConfigComponent } from '@features/events/event-config/event-config.component';
 import { EventEditComponent } from '@features/events/event-edit/event-edit.component';
 import { EventCreationComponent } from '@features/events/event-creation/event-creation.component';
 import { EventWikiComponent } from '@features/events/event-wiki/event-wiki.component';
@@ -33,8 +33,8 @@ const routes: Routes = [
     component: EventEditComponent,
   },
   {
-    path: ':id/admin',
-    component: EventAdminComponent,
+    path: ':id/config',
+    component: EventConfigComponent,
   },
   {
     path: ':id/wiki',